+++ espresso-macchiato/src/main/java/de/nenick/espressomacchiato/elements/support/EspRecyclerViewLayoutManager.java
@@ -1,11 +1,14 @@
 package de.nenick.espressomacchiato.elements.support;
 
 import android.support.test.espresso.ViewInteraction;
+import android.support.v7.widget.GridLayoutManager;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 
 import de.nenick.espressomacchiato.assertions.support.GridLayoutManagerColumnCountAssertion;
-import de.nenick.espressomacchiato.assertions.support.LayoutManagerItemVisibilityAssertion;
 
 import static de.nenick.espressomacchiato.assertions.support.LayoutManagerTypeAssertion.isGridLayoutManager;
+import static de.nenick.espressomacchiato.assertions.support.LayoutManagerTypeAssertion.isLinearLayoutManager;
 
 /**
  * Assertions for RecyclerView layout manager.
@@ -16,16 +19,43 @@ public class EspRecyclerViewLayoutManager {
 
     private final ViewInteraction recyclerViewInteraction;
 
+    /**
+     * Create new element instance.
+     *
+     * @param recyclerViewInteraction prepared interaction for target {@link RecyclerView}
+     * @since Espresso Macchiato 0.6
+     */
     public EspRecyclerViewLayoutManager(ViewInteraction recyclerViewInteraction) {
         this.recyclerViewInteraction = recyclerViewInteraction;
     }
 
+    /**
+     * Check that the layout manager is of type {@link LinearLayoutManager}
+     *
+     * @since Espresso Macchiato 0.6
+     */
+    public void assertLinearLayoutManager() {
+        recyclerViewInteraction.check(isLinearLayoutManager());
+    }
+
+    /**
+     * Check that the layout manager is of type {@link GridLayoutManager}
+     *
+     * @since Espresso Macchiato 0.6
+     */
     public void assertGridLayoutManager() {
         recyclerViewInteraction.check(isGridLayoutManager());
     }
 
+    /**
+     * Check that the {@link GridLayoutManager} has expected count of column.
+     *
+     * @param expectedColumnCount expected count of columns
+     * @since Espresso Macchiato 0.6
+     */
     public void assertColumnCount(int expectedColumnCount) {
         assertGridLayoutManager();
         recyclerViewInteraction.check(new GridLayoutManagerColumnCountAssertion(expectedColumnCount));
     }
+
 }