+++ espresso-macchiato/src/main/java/de/nenick/espressomacchiato/matchers/EspAllOfBuilder.java
@@ -13,35 +13,81 @@ import java.util.ArrayList;
 
 import de.nenick.espressomacchiato.elements.EspView;
 
+/**
+ * Fluent builder fo allOf matcher.
+ *
+ * @param <ElementT> View element type which will be used to perform actions and assertions on match.
+ * @since Espresso Macchiato 0.4
+ */
 public abstract class EspAllOfBuilder<ElementT extends EspView> {
 
     protected ArrayList<Matcher<? super View>> matcher = Lists.newArrayList();
 
+    /**
+     * Add matcher for element id.
+     *
+     * @param resourceId expected id
+     * @return The current builder.
+     * @since Espresso Macchiato 0.4
+     */
     public EspAllOfBuilder<ElementT> withId(int resourceId) {
         matcher.add(ViewMatchers.withId(resourceId));
         return this;
     }
 
+    /**
+     * Add matcher that element is displayed.
+     *
+     * @return The current builder.
+     * @since Espresso Macchiato 0.4
+     */
     public EspAllOfBuilder<ElementT> withIsDisplayed() {
         matcher.add(ViewMatchers.isDisplayed());
         return this;
     }
 
+    /**
+     * Add matcher that element must have an expected parent with id.
+     *
+     * @param resourceId expected parent id
+     * @return The current builder.
+     * @since Espresso Macchiato 0.4
+     */
     public EspAllOfBuilder<ElementT> withParentInHierarchy(int resourceId) {
         matcher.add(ViewMatchers.isDescendantOfA(ViewMatchers.withId(resourceId)));
         return this;
     }
 
+    /**
+     * Add matcher for text.
+     *
+     * @param text expected text
+     * @return The current builder.
+     * @since Espresso Macchiato 0.6
+     */
     public EspAllOfBuilder<ElementT> withText(String text) {
         matcher.add(ViewMatchers.withText(text));
         return this;
     }
 
+    /**
+     * Add matcher for visibility.
+     *
+     * @param visible expected visibility
+     * @return The current builder.
+     * @since Espresso Macchiato 0.6
+     */
     public EspAllOfBuilder<ElementT> withVisibility(ViewMatchers.Visibility visible) {
         matcher.add(ViewMatchers.withEffectiveVisibility(visible));
         return this;
     }
 
+    /**
+     * Create an view element instance with the given matcher.
+     *
+     * @return New instance of the defined view element.
+     * @since Espresso Macchiato 0.4
+     */
     public ElementT build() {
         try {
             Matcher<View> allOfMatcher = Matchers.allOf(matcher);
@@ -51,7 +97,7 @@ public abstract class EspAllOfBuilder<ElementT extends EspView> {
         }
     }
 
-    protected Class<ElementT> getGenericClass() {
+    private Class<ElementT> getGenericClass() {
         Type genericSuperclass = getClass().getGenericSuperclass();
         if (genericSuperclass instanceof ParameterizedType) {
             //noinspection unchecked