+++ espresso-macchiato/src/main/java/de/nenick/espressomacchiato/tools/EspAppDataTool.java
@@ -21,11 +21,15 @@ import static org.junit.Assert.assertTrue;
  * new ActivityTestRule(Activity.class, false, false)<br>
  * clearApplicationData()<br>
  * ActivityTestRule.lunchActivity()
+ *
+ * @since Espresso Macchiato 0.3
  */
 public class EspAppDataTool {
 
     /**
      * Clear all application data except screenshots taken with {@link EspScreenshotTool}.
+     *
+     * @since Espresso Macchiato 0.3
      */
     public static void clearApplicationData() {
         clearStorageExceptScreenshots();
@@ -36,11 +40,13 @@ public class EspAppDataTool {
 
     /**
      * Clear all shared preferences.
+     *
+     * @since Espresso Macchiato 0.3
      */
     public static void clearSharedPreferences() {
         String[] sharedPreferencesFileNames = getSharedPreferencesFilesLocation().list();
 
-        if(sharedPreferencesFileNames == null) {
+        if (sharedPreferencesFileNames == null) {
             // occurs when "shared_prefs" folder was not created yet
             return;
         }
@@ -60,7 +66,12 @@ public class EspAppDataTool {
     }
 
     /**
-     * only works if all database connections are closed. does not produce error if connection still open.
+     * Clear all databases.
+     *
+     * Only works if all database connections are closed.
+     * Does not produce error if connection still open.
+     *
+     * @since Espresso Macchiato 0.3
      */
     public static void clearDatabase() {
         String[] databaseList = InstrumentationRegistry.getTargetContext().databaseList();
@@ -68,27 +79,41 @@ public class EspAppDataTool {
 
             // when transaction rollback files exists they are always locked so we can't delete them
             if (database.contains(".db-journal")) {
+                InstrumentationRegistry.getTargetContext().deleteDatabase(database);
                 continue;
             }
 
-            // not exist but listed db files (occurs with web views)
+            // when using transaction write ahead logging then this db files are listed but often they don't exist
             if (database.contains(".db-wal") || database.contains(".db-shm")) {
+                InstrumentationRegistry.getTargetContext().deleteDatabase(database);
                 continue;
             }
 
             Log.v("EspressoMacchiato", "deleting " + database);
 
-            assertThat(InstrumentationRegistry.getTargetContext().getDatabasePath(database).exists(), is(true));
+            File databasePath = InstrumentationRegistry.getTargetContext().getDatabasePath(database);
+            assertThat("db exist at " + databasePath, databasePath.exists(), is(true));
             assertThat("could not delete " + database, InstrumentationRegistry.getTargetContext().deleteDatabase(database), is(true));
-            assertThat(InstrumentationRegistry.getTargetContext().getDatabasePath(database).exists(), is(false));
+            assertThat("db exist not anymore at " + databasePath, databasePath.exists(), is(false));
         }
     }
 
+    /**
+     * Clear cached files.
+     *
+     * @since Espresso Macchiato 0.3
+     */
     public static void clearCache() {
         File cacheDir = InstrumentationRegistry.getTargetContext().getCacheDir();
         assertThat(deleteRecursive(cacheDir), is(true));
     }
 
+    /**
+     * Clear stored files.
+     *
+     * @param excludes directory names which will be skipped
+     * @since Espresso Macchiato 0.3
+     */
     public static void clearStorage(String... excludes) {
         File filesDir = InstrumentationRegistry.getTargetContext().getFilesDir();
         String[] directoryContent = filesDir.list();
@@ -98,6 +123,11 @@ public class EspAppDataTool {
 
     }
 
+    /**
+     * Clear stored files except test screenshots folder.
+     *
+     * @since Espresso Macchiato 0.3
+     */
     public static void clearStorageExceptScreenshots() {
         clearStorage(EspScreenshotTool.screenshotFolderName);
     }